option to sim_fsm2oshd for forest only

+++ TopoPyScale/sim_fsm2oshd.py
@@ -68,7 +68,7 @@ def fsm2oshd_sim_parallel(simulation_path='./fsm_sim,',
 
     Args:
         fsm_forest_nam (str): file pattern of namelist file for forest setup. Default = 'fsm_sim/fsm__forest*.nam'
-        fsm_open_nam (str): file pattern of namelist file for open setup. Default = 'fsm_sim/fsm__open*.nam'
+        fsm_open_nam (str): file pattern of namelist file for open setup. Default = 'fsm_sim/fsm__open*.nam'. If None, then only performs forest simulation
         fsm_exec (str): path to FSM executable. Default = './FSM'
         n_cores (int): number of cores. Default = 6
         delete_nlst_files (bool): delete simulation configuration files or not. Default=True
@@ -84,19 +84,22 @@ def fsm2oshd_sim_parallel(simulation_path='./fsm_sim,',
     tu.multicore_pooling(_run_fsm2oshd, fun_param, n_cores)
     print('---> FSM2oshd forest simulations finished.')
 
-    # 2. execute FSM on multicore for Open
-    nlst_open = glob.glob(str(p / fsm_open_nam))
-    fun_param = zip([fsm_exec]*len(nlst_open), nlst_open)
-    tu.multicore_pooling(_run_fsm2oshd, fun_param, n_cores)
-    print('---> FSM2oshd open simulations finished.')
+    if fsm_open_nam is not None:
+
+        # 2. execute FSM on multicore for Open
+        nlst_open = glob.glob(str(p / fsm_open_nam))
+        fun_param = zip([fsm_exec]*len(nlst_open), nlst_open)
+        tu.multicore_pooling(_run_fsm2oshd, fun_param, n_cores)
+        print('---> FSM2oshd open simulations finished.')
 
     # 3. remove nlist files
     if delete_nlst_files:
         print('---> Removing FSM simulation config files')
         for file in nlst_forest:
             os.remove(file)
-        for file in nlst_open:
-            os.remove(file)
+        if fsm_open_nam is not None:
+            for file in nlst_open:
+                os.remove(file)
